[REFACTOR] FORM 컴포넌트 분리 후 리팩토링
